big refactor because the "refactor" file was too long by a lot
new file:   TODO

big refactor.  break out into smaller files.  untested.
gearing up for a slight modification of function for freeways,
so that I can store detector summations by hour in pgdb
 modified:   lib/detector.area.data.refactor.js
 new file:   lib/doDataQuery.js
 new file:   lib/doneDataSumUp.js
 new file:   lib/doneGeoQuery.js
 new file:   lib/geoQuery.js

This is the file with the beginnings of how to save to db.
It is set to dump to csv, which can then be loaded into a table.
 new file:   lib/doneDataFeatures.js

fixes.
 modified:   lib/aggregate_from_hourly.js
require underscore
 modified:   package.json